update wrapper tests for new ExpandableWrapper object
